github-insights/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
load-dotenv==0.1.0
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
